github_copyfile/lambda_docker/requirements.txt
torch
torchvision
Pillow
boto3